Add increase_native_token_supply example
