+++ examples/12_increase_native_token_supply.rs
@@ -1,15 +1,15 @@
 // Copyright 2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-//! cargo run --example 13_burn_nft --release
-// In this example we will burn an existing nft output
+//! cargo run --example increase_native_token_supply --release
+// In this example we will melt an existing native token with its foundry
 // Rename `.env.example` to `.env` first
 
 use std::{env, str::FromStr};
 
 use dotenv::dotenv;
-use iota_client::block::output::NftId;
-use iota_wallet::{account_manager::AccountManager, Result};
+use iota_client::block::output::TokenId;
+use iota_wallet::{account_manager::AccountManager, Result, U256};
 
 #[tokio::main]
 async fn main() -> Result<()> {
@@ -23,29 +23,35 @@ async fn main() -> Result<()> {
     let account = manager.get_account("Alice").await?;
 
     let balance = account.balance().await?;
-    println!("Balance before burning:\n{balance:?}",);
+    println!("Balance before melting:\n{balance:?}",);
 
     // Set the stronghold password
     manager
         .set_stronghold_password(&env::var("STRONGHOLD_PASSWORD").unwrap())
         .await?;
 
-    // Replace with an NftId that is available in the account
-    let nft_id = NftId::from_str("0xe192461b30098a5da889ef6abc9e8130bf3b2d980450fa9201e5df404121b932")?;
-    let transaction = account.burn_nft(nft_id, None).await?;
+    // Replace with a TokenId that is available in the account, the foundry output which minted it, also needs to be
+    // available.
+    let token_id = TokenId::from_str("0x08847bd287c912fadedb6bf38900bda9f2d377b75b2a0bece8738699f56ebca4130100000000")?;
 
-    let _ = match transaction.block_id {
+    // Mint some more native tokens
+    let mint_amount = U256::from(10);
+    let mint_transaction = account
+        .increase_native_token_supply(token_id, mint_amount, None, None)
+        .await?;
+
+    let _ = match mint_transaction.transaction.block_id {
         Some(block_id) => account.retry_until_included(&block_id, None, None).await?,
         None => {
             return Err(iota_wallet::Error::BurningOrMeltingFailed(
-                "burn nft failed to submitted".to_string(),
+                "mint native token transaction failed to submitted".to_string(),
             ));
         }
     };
 
     let balance = account.sync(None).await?;
 
-    println!("Balance after burning:\n{balance:?}",);
+    println!("Balance after minting:\n{balance:?}",);
 
     Ok(())
 }
